fix excel bug

+++ database/migrations/2020_04_20_082920_create_order_items_table.php
@@ -21,8 +21,8 @@ class CreateOrderItemsTable extends Migration
             $table->unsignedInteger('product_relation_id');
             $table->string('spec_name')->nullable();
 
-            $table->decimal('quantity');
-            $table->decimal('price');
+            $table->unsignedInteger('quantity');
+            $table->unsignedInteger('price');
             $table->tinyInteger('status')->default(0);
             $table->timestamp('create_date')->nullable();
             $table->timestamp('update_date')->nullable();
